+++ CHANGELOG.md
@@ -1,3 +1,3 @@
 > Read detailed change logs in https://github.com/sakurawald/fuji/commits/dev/
   
-- [world.border] fix: ensure the `client-side world border` is synced with the server side.
+- [world.border] fix: ensure the `client-side world border` is synced after teleport.